--- src/components/Client/ClientCard.css
@@ -1,3 +1,0 @@
-.icon{
-    font-size: 100px;
-}
