+++ src/pages/Clients/ClientDetails.tsx
@@ -1,12 +1,45 @@
-import { IonBackButton, IonButton, IonButtons, IonContent, IonHeader, IonIcon, IonItem, IonPage, IonText, IonTitle, IonToolbar, useIonAlert } from "@ionic/react";
+import { 
+    IonBackButton, 
+    IonButton, 
+    IonButtons, 
+    IonCard, 
+    IonCardContent, 
+    IonCardHeader, 
+    IonContent, 
+    IonHeader, 
+    IonIcon, 
+    IonItem, 
+    IonPage, 
+    IonText, 
+    IonTitle, 
+    IonToolbar, 
+    useIonAlert 
+} from "@ionic/react";
 import { useEffect, useState } from "react";
 import { useParams } from "react-router";
-import ClientCard from "../../components/Client/ClientCard";
 import IClientData from "../../types/client.type";
 import ClientDataService from "../../services/client.service"
 import './ClientDetails.css';
-import { pencilOutline, pencilSharp, trashBinOutline, trashBinSharp } from "ionicons/icons";
+import { 
+    calendarClearOutline, 
+    calendarClearSharp, 
+    fileTrayFullOutline, 
+    fileTrayFullSharp, 
+    giftOutline, 
+    giftSharp, 
+    pencilOutline, 
+    pencilSharp, 
+    personOutline, 
+    personSharp, 
+    pinOutline, 
+    pinSharp, 
+    trashBinOutline, 
+    trashBinSharp 
+} from "ionicons/icons";
 import EditClient from "../../components/Client/EditClient";
+import ILawyercase from "../../types/lawyercase.type";
+import {format, parseISO} from 'date-fns';
+
 
 interface ParamsInterface{
     id: string;
@@ -41,6 +74,9 @@ const ClientDetails = () => {
         setSelectedClient(client)
         setIsEdit(true)
     }
+    const formatDate = (value: string) => {
+        return format(parseISO(value), 'dd/MM/yyyy');
+    };   
 
     useEffect(() => {
         ClientDataService.get(params.id)
@@ -92,11 +128,67 @@ const ClientDetails = () => {
                 </IonItem>
             </IonHeader>
             <IonContent>
-                {client ? 
-                <ClientCard client={client}/> : 
-                <IonItem>
-                    <IonText color="danger">Ce client n'existe pas!</IonText>
-                </IonItem>
+                {client ?
+                    <>    
+                        <IonCard>
+                            <IonCardContent>            
+                                <IonItem lines="none">
+                                    <IonIcon ios={personOutline} md={personSharp}/>
+                                    <IonTitle>{client.name + ' ' + client.firstname}</IonTitle>
+                                </IonItem>
+                                    {client.createdAt ? (
+                                        <IonItem lines="none">
+                                            <IonIcon ios={calendarClearOutline} md={calendarClearSharp}/>
+                                            <IonTitle>Client depuis le {formatDate(client.createdAt)}</IonTitle>
+                                        </IonItem>
+                                    ) : null}
+                                    {client.birthdate ? (
+                                        <IonItem lines="none">
+                                            <IonIcon ios={giftOutline} md={giftSharp}/>
+                                            <IonTitle>{client.birthdate}</IonTitle>
+                                        </IonItem>
+                                    ): null}                    
+                            </IonCardContent>
+                        </IonCard>
+            
+                        <IonCard>
+                            <IonCardHeader>
+                                <IonItem lines="none">
+                                    <IonIcon ios={pinOutline} md={pinSharp}/>
+                                    <IonTitle>Adresse</IonTitle>
+                                </IonItem>
+                            </IonCardHeader>
+            
+                            <IonCardContent>
+                                <IonText>
+                                    <h2>{client.address}</h2>
+                                </IonText>
+                            </IonCardContent>
+                        </IonCard>
+            
+                        <IonCard>
+                            <IonCardHeader>
+                                <IonItem lines="none">
+                                    <IonIcon ios={fileTrayFullOutline} md={fileTrayFullSharp}/>
+                                    <IonTitle>Dossiers</IonTitle>
+                                </IonItem>
+                            </IonCardHeader>
+            
+                            <IonCardContent>
+                                {client.cases?.map((lawyercase: ILawyercase, index: number) => (
+                                    <IonItem lines="none" key={index}>                            
+                                        <IonText>{lawyercase.ref}</IonText>
+                                    </IonItem>
+                                ))} 
+                            </IonCardContent>
+                        </IonCard>
+                    </>           
+                    
+                
+                 : 
+                    <IonItem>
+                        <IonText color="danger">Ce client n'existe pas!</IonText>
+                    </IonItem>
                 }
             </IonContent>
