+++ src/pages/Records/RecordDetails.tsx
@@ -1,12 +1,32 @@
-import { IonBackButton, IonButton, IonButtons,  IonContent,  IonHeader, IonIcon, IonItem,  IonPage, IonText, IonTitle, IonToolbar, useIonAlert } from '@ionic/react';
-import {   pencilOutline, pencilSharp, trashBinOutline, trashBinSharp } from 'ionicons/icons';
+import { 
+    IonBackButton, 
+    IonButton, 
+    IonButtons,  
+    IonContent,  
+    IonHeader, 
+    IonIcon, 
+    IonItem,  
+    IonPage, 
+    IonText, 
+    IonTitle, 
+    IonToolbar, 
+    useIonAlert,
+    IonCard, 
+    IonCardContent, 
+    IonCardHeader, 
+    IonLabel,     
+} from '@ionic/react';
+import {   addOutline, addSharp, alertOutline, alertSharp, calendarClearOutline, calendarClearSharp, ellipse, fileTrayFullOutline, fileTrayFullSharp, fileTrayOutline, pencilOutline, pencilSharp, personAddOutline, personAddSharp, personOutline, personSharp, timeOutline, timeSharp, trashBinOutline, trashBinSharp } from 'ionicons/icons';
 import React, { useEffect, useState } from 'react';
 import { useParams } from 'react-router';
 import ILawyercase from '../../types/lawyercase.type';
 import LawyercaseDataService from "../../services/lawyercase.service"
 import './Records.css';
-import RecordCard from '../../components/Dossier/RecordCard';
 import EditRecord from '../../components/Dossier/EditRecord';
+import AddEvent from '../../components/Évènement/AddEvent';
+import AddClientToCaseModal from '../../components/Dossier/AddClientToCase';
+import {format, parseISO} from 'date-fns';
+import lawyercaseService from '../../services/lawyercase.service';
 
 const RecordDetails: React.FC = () =>
 {
@@ -17,8 +37,11 @@ const RecordDetails: React.FC = () =>
     const params = useParams<ParamsInterface>();
     const [record, setRecord] = useState<ILawyercase>();
     const [isEdit, setIsEdit] = useState(false);
-    const [present] = useIonAlert();
+    const [isOpen, setIsOpen] = useState(false);
+    const [Delete, setDelete] = useState(false);
+    const [addClientModal, setAddClientModal] = useState(false)
     const [selectedRecord, setSelectedRecord] = useState<ILawyercase>();
+    const [present] = useIonAlert();
 
 
     const deleteRecord = (id:string) => {
@@ -40,16 +63,26 @@ const RecordDetails: React.FC = () =>
         setIsEdit(true)
     }
 
+    const formatDate = (value: string) => {
+        return format(parseISO(value), 'dd/MM/yyyy');
+    };
+
+    const handleDeleteClient = (clientId: string) => {
+        setDelete(true);
+        lawyercaseService.removeClient(record?.id, clientId)
+    }   
+
     useEffect(() => {
         LawyercaseDataService.get(params.id)
             .then((response: any) => {
-                console.log(response.data);
                 setRecord(response.data);
             })
             .catch((e: Error) => {
                 console.log(e);
             });
-    }, [params.id, isEdit]);
+        setDelete(false);
+        
+    }, [params.id, isEdit, Delete, isOpen, addClientModal, present]);
     
      return (
         <IonPage>
@@ -90,7 +123,118 @@ const RecordDetails: React.FC = () =>
             </IonHeader>
             <IonContent>
                 {record ? 
-                <RecordCard record={record}/> : 
+                 <>
+                 {record ? (
+                     <>    
+                     <IonCard>
+                         <IonCardContent>
+                             <IonItem lines="none">
+                                     <IonIcon />
+                                     <IonTitle>{record.ref}</IonTitle>
+                             </IonItem>
+                             <IonItem lines="none">
+                                 <IonIcon color={record.closed_at ? "danger" : "success"} ios={ellipse} md={ellipse} />
+                                 <IonTitle>{record.closed_at ? "Clôturé" : "En cours"}</IonTitle>
+                             </IonItem>
+                             {record.createdAt ? (
+                                 <IonItem lines="none">
+                                         <IonIcon ios={calendarClearOutline} md={calendarClearSharp}/>
+                                         <IonTitle>{formatDate(record.createdAt)}</IonTitle>
+                                 </IonItem>
+                             ) : null}
+                         </IonCardContent>
+                     </IonCard>
+         
+                     <IonCard>
+                         <IonCardHeader>
+                             <IonItem lines="none">
+                                 <IonIcon ios={fileTrayFullOutline} md={fileTrayFullSharp}/>
+                                 <IonTitle>Description</IonTitle>
+                             </IonItem>
+                         </IonCardHeader>
+         
+                         <IonCardContent>
+                             <IonText>
+                                 <h2>{record.description}</h2>
+                             </IonText>
+                         </IonCardContent>
+                     </IonCard>
+         
+                     <IonCard>
+                         <IonCardHeader>
+                             <IonItem lines="none">
+                                 <IonIcon ios={personOutline} md={personSharp}/>
+                                 <IonTitle>Clients concernés</IonTitle>
+                                 <IonButtons slot='end'>
+                                     <IonButton color='primary' onClick={() => {setAddClientModal(true)}} >
+                                         <IonIcon color="primary" ios={personAddOutline} md={personAddSharp} />
+                                     </IonButton>
+                                 </IonButtons>
+                             </IonItem>
+                         </IonCardHeader>
+         
+                         <IonCardContent>
+                             {record.clients?.map((client, index) => (
+                                 <IonItem lines="none" key={index}>   
+                                     <IonItem lines="none">
+                                         <IonButton color="danger" onClick={() => {
+                                             present({
+                                                 cssClass: 'my-css',
+                                                 header: 'Suppression d\'un client',
+                                                 message: 'êtes-vous sûr de vouloir supprimer ce client ?',
+                                                 buttons: [
+                                                   {text: 'Annuler', role: 'cancel'},
+                                                   { text: 'Confirmer', handler: () => handleDeleteClient(client.id)}
+                                                 ],                        
+                                               }) 
+                                             }}>
+                                             <IonIcon ios={trashBinOutline} md={trashBinSharp} />
+                                         </IonButton>
+                                     </IonItem>
+                                     <IonLabel>{client.name + ' ' + client.firstname}</IonLabel>
+                                 </IonItem>
+                             ))}                    
+                         </IonCardContent>
+                     </IonCard>
+         
+         
+                     <IonCard>
+                         <IonCardHeader>
+                             <IonItem lines="none">
+                                 <IonIcon ios={alertOutline} md={alertSharp}/>
+                                 <IonTitle>Évènements</IonTitle>
+                                 <IonButtons slot='end'>
+                                     <IonButton color='primary' onClick={() => {setIsOpen(true)}} >
+                                         <IonIcon color="primary" ios={addOutline} md={addSharp} />
+                                     </IonButton>
+                                 </IonButtons>
+                             </IonItem>
+                             <IonItem lines='none'>
+                             </IonItem>
+                         </IonCardHeader>
+         
+                         <IonCardContent>
+                             <IonText>
+                             </IonText>
+                         </IonCardContent>
+                     </IonCard>
+                     
+                     <IonCard>
+                         <IonCardHeader>
+                             <IonItem lines="none">
+                                 <IonIcon ios={timeOutline} md={timeSharp}/>
+                                 <IonTitle>Total : Temps en heures</IonTitle>
+                             </IonItem>
+                         </IonCardHeader>
+         
+                         <IonCardContent>
+                             
+                         </IonCardContent>
+                     </IonCard>                     
+                     </>
+                 ) : null}                     
+                 </>
+                 : 
                 <IonItem>
                     <IonText color="danger">Ce dossier n'existe pas!</IonText>
                 </IonItem>
@@ -103,6 +247,10 @@ const RecordDetails: React.FC = () =>
                     setIsOpen={() => setIsEdit(false)}
                 />
             ) : null}
+            {record ? (
+                <AddClientToCaseModal recordClients={record.clients ?? []} record={record} isOpen={addClientModal} setIsOpen={setAddClientModal}/>
+            )  : null}
+            <AddEvent  isOpen={isOpen} setIsOpen={() => setIsOpen(false)}/>
         </IonPage>
     );
 }
